+++ __tests__/lint/assertions-non-empty-off/snapshot.js
@@ -1,6 +1,6 @@
 // Jest Snapshot v1, https://goo.gl/fbAQLP
 
-exports[`E2E lint no-servers-empty-enum 1`] = `
+exports[`E2E lint assertions-non-empty-off 1`] = `
 
 validating /openapi.yaml...
 /openapi.yaml: validated in <test>ms