+++ packages/cli/src/commands/split/index.ts
@@ -1,6 +1,7 @@
 import { red, blue, yellow, green } from 'colorette';
 import * as fs from 'fs';
-import { parseYaml, slash, isRef } from '@redocly/openapi-core';
+import { parseYaml, slash, isRef, isTruthy } from '@redocly/openapi-core';
+import type { OasRef } from '@redocly/openapi-core';
 import * as path from 'path';
 import { performance } from 'perf_hooks';
 const isEqual = require('lodash.isequal');
@@ -33,17 +34,17 @@ import {
   Referenced,
 } from './types';
 
-export async function handleSplit(argv: { entrypoint: string; outDir: string; separator: string }) {
+export async function handleSplit(argv: { api: string; outDir: string; separator: string }) {
   const startedAt = performance.now();
-  const { entrypoint, outDir, separator } = argv;
-  validateDefinitionFileName(entrypoint!);
-  const openapi = readYaml(entrypoint!) as Oas3Definition | Oas3_1Definition;
+  const { api, outDir, separator } = argv;
+  validateDefinitionFileName(api!);
+  const openapi = readYaml(api!) as Oas3Definition | Oas3_1Definition;
   splitDefinition(openapi, outDir, separator);
   process.stderr.write(
-    `🪓 Document: ${blue(entrypoint!)} ${green('is successfully split')}
+    `🪓 Document: ${blue(api!)} ${green('is successfully split')}
     and all related files are saved to the directory: ${blue(outDir)} \n`
   );
-  printExecutionTime('split', startedAt, entrypoint!);
+  printExecutionTime('split', startedAt, api!);
 }
 
 function splitDefinition(
@@ -267,11 +268,11 @@ function gatherComponentsFiles(
   componentName: string,
   filename: string
 ) {
-  let inherits = [];
+  let inherits: string[] = [];
   if (componentType === OPENAPI3_COMPONENT.Schemas) {
     inherits = ((components?.[componentType]?.[componentName] as Oas3Schema)?.allOf || [])
-      .map((s: any) => s.$ref)
-      .filter(Boolean);
+      .map(({ $ref }) => $ref)
+      .filter(isTruthy);
   }
   componentsFiles[componentType] = componentsFiles[componentType] || {};
   componentsFiles[componentType][componentName] = { inherits, filename };
@@ -301,7 +302,7 @@ function iteratePathItems(
         continue;
       }
       for (const sample of methodDataXCode) {
-        if (sample.source && (sample.source as any).$ref) continue;
+        if (sample.source && (sample.source as unknown as OasRef).$ref) continue;
         const sampleFileName = path.join(
           openapiDir,
           'code_samples',
@@ -312,6 +313,7 @@ function iteratePathItems(
 
         fs.mkdirSync(path.dirname(sampleFileName), { recursive: true });
         fs.writeFileSync(sampleFileName, sample.source);
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         sample.source = {
           $ref: slash(path.relative(outDir, sampleFileName)),
@@ -340,6 +342,7 @@ function iterateComponents(
     componentTypes.forEach(iterateAndGatherComponentsFiles);
     componentTypes.forEach(iterateComponentTypes);
 
+    // eslint-disable-next-line no-inner-declarations
     function iterateAndGatherComponentsFiles(componentType: Oas3ComponentName) {
       const componentDirPath = path.join(componentsDir, componentType);
       for (const componentName of Object.keys(components?.[componentType] || {})) {
@@ -348,6 +351,7 @@ function iterateComponents(
       }
     }
 
+    // eslint-disable-next-line no-inner-declarations
     function iterateComponentTypes(componentType: Oas3ComponentName) {
       const componentDirPath = path.join(componentsDir, componentType);
       createComponentDir(componentDirPath, componentType);