+++ packages/cli/src/index.ts
@@ -18,10 +18,10 @@ yargs
   .version('version', 'Show version number.', version)
   .help('help', 'Show help.')
   .command(
-    'stats [entrypoint]',
+    'stats [api]',
     'Gathering statistics for a document.',
     (yargs) =>
-      yargs.positional('entrypoint', { type: 'string' }).option({
+      yargs.positional('api', { type: 'string' }).option({
         config: { description: 'Specify path to the config file.', type: 'string' },
         format: {
           description: 'Use a specific output format.',
@@ -29,14 +29,17 @@ yargs
           default: 'stylish' as OutputFormat,
         },
       }),
-    handleStats
+    (argv) => {
+      process.env.REDOCLY_CLI_COMMAND = 'stats';
+      handleStats(argv);
+    }
   )
   .command(
-    'split [entrypoint]',
+    'split [api]',
     'Split definition into a multi-file structure.',
     (yargs) =>
       yargs
-        .positional('entrypoint', {
+        .positional('api', {
           description: 'API definition file that you want to split',
           type: 'string',
         })
@@ -53,15 +56,18 @@ yargs
             default: '_',
           },
         })
-        .demandOption('entrypoint'),
-    handleSplit
+        .demandOption('api'),
+    (argv) => {
+      process.env.REDOCLY_CLI_COMMAND = 'split';
+      handleSplit(argv);
+    }
   )
   .command(
-    'join [entrypoints...]',
+    'join [apis...]',
     'Join definitions [experimental].',
     (yargs) =>
       yargs
-        .positional('entrypoints', {
+        .positional('apis', {
           array: true,
           type: 'string',
           demandOption: true,
@@ -89,15 +95,16 @@ yargs
           },
         }),
     (argv) => {
+      process.env.REDOCLY_CLI_COMMAND = 'join';
       handleJoin(argv, version);
     }
   )
   .command(
-    'push [maybeEntrypointOrAliasOrDestination] [maybeDestination] [maybeBranchName]',
+    'push [maybeApiOrDestination] [maybeDestination] [maybeBranchName]',
     'Push an API definition to the Redocly API registry.',
     (yargs) =>
       yargs
-        .positional('maybeEntrypointOrAliasOrDestination', { type: 'string' })
+        .positional('maybeApiOrDestination', { type: 'string' })
         .positional('maybeDestination', { type: 'string' })
         .positional('maybeBranchName', { type: 'string' })
         .option({
@@ -127,13 +134,16 @@ yargs
         })
         .implies('batch-id', 'batch-size')
         .implies('batch-size', 'batch-id'),
-    transformPush(handlePush)
+    (argv) => {
+      process.env.REDOCLY_CLI_COMMAND = 'push';
+      transformPush(handlePush)(argv);
+    }
   )
   .command(
-    'lint [entrypoints...]',
+    'lint [apis...]',
     'Lint definition.',
     (yargs) =>
-      yargs.positional('entrypoints', { array: true, type: 'string', demandOption: true }).option({
+      yargs.positional('apis', { array: true, type: 'string', demandOption: true }).option({
         format: {
           description: 'Use a specific output format.',
           choices: [
@@ -183,14 +193,15 @@ yargs
         },
       }),
     (argv) => {
+      process.env.REDOCLY_CLI_COMMAND = 'lint';
       handleLint(argv, version);
     }
   )
   .command(
-    'bundle [entrypoints...]',
+    'bundle [apis...]',
     'Bundle definition.',
     (yargs) =>
-      yargs.positional('entrypoints', { array: true, type: 'string', demandOption: true }).options({
+      yargs.positional('apis', { array: true, type: 'string', demandOption: true }).options({
         output: { type: 'string', alias: 'o' },
         format: {
           description: 'Use a specific output format.',
@@ -259,6 +270,7 @@ yargs
         },
       }),
     (argv) => {
+      process.env.REDOCLY_CLI_COMMAND = 'bundle';
       handleBundle(argv, version);
     }
   )
@@ -277,23 +289,27 @@ yargs
           choices: regionChoices,
         },
       }),
-    handleLogin
+    (argv) => {
+      process.env.REDOCLY_CLI_COMMAND = 'login';
+      handleLogin(argv);
+    }
   )
   .command(
     'logout',
     'Clear your stored credentials for the Redocly API registry.',
     (yargs) => yargs,
     async () => {
+      process.env.REDOCLY_CLI_COMMAND = 'logout';
       const client = new RedoclyClient();
       client.logout();
       process.stdout.write('Logged out from the Redocly account. ✋\n');
     }
   )
   .command(
-    'preview-docs [entrypoint]',
+    'preview-docs [api]',
     'Preview API reference docs for the specified definition.',
     (yargs) =>
-      yargs.positional('entrypoint', { type: 'string' }).options({
+      yargs.positional('api', { type: 'string' }).options({
         port: {
           alias: 'p',
           type: 'number',
@@ -330,7 +346,10 @@ yargs
           type: 'string',
         },
       }),
-    previewDocs
+    (argv) => {
+      process.env.REDOCLY_CLI_COMMAND = 'preview-docs';
+      previewDocs(argv);
+    }
   )
   .completion('completion', 'Generate completion script.')
   .demandCommand(1)