+++ packages/core/src/redocly/registry-api.ts
@@ -1,10 +1,17 @@
 import fetch, { RequestInit, HeadersInit } from 'node-fetch';
-import { RegistryApiTypes } from './registry-api-types';
+import type {
+  NotFoundProblemResponse,
+  PrepareFileuploadOKResponse,
+  PrepareFileuploadParams,
+  PushApiParams,
+} from './registry-api-types';
+import type { AccessTokens, Region } from '../config/types';
 import { DEFAULT_REGION, DOMAINS } from '../config/config';
 import { isNotEmptyObject } from '../utils';
 const version = require('../../package.json').version;
 
-import type { AccessTokens, Region } from '../config/types';
+export const currentCommand =
+  typeof process !== 'undefined' ? process.env?.REDOCLY_CLI_COMMAND || '' : '';
 
 export class RegistryApi {
   constructor(private accessTokens: AccessTokens, private region: Region) {}
@@ -23,7 +30,10 @@ export class RegistryApi {
   }
 
   private async request(path = '', options: RequestInit = {}, region?: Region) {
-    const headers = Object.assign({}, options.headers || {}, { 'x-redocly-cli-version': version });
+    const headers = Object.assign({}, options.headers || {}, {
+      'x-redocly-cli-version': version,
+      'user-agent': `redocly-cli / ${version} ${currentCommand}`,
+    });
 
     if (!headers.hasOwnProperty('authorization')) {
       throw new Error('Unauthorized');
@@ -39,7 +49,7 @@ export class RegistryApi {
     }
 
     if (response.status === 404) {
-      const body: RegistryApiTypes.NotFoundProblemResponse = await response.json();
+      const body: NotFoundProblemResponse = await response.json();
       throw new Error(body.code);
     }
 
@@ -71,7 +81,7 @@ export class RegistryApi {
     filesHash,
     filename,
     isUpsert,
-  }: RegistryApiTypes.PrepareFileuploadParams): Promise<RegistryApiTypes.PrepareFileuploadOKResponse> {
+  }: PrepareFileuploadParams): Promise<PrepareFileuploadOKResponse> {
     const response = await this.request(
       `/${organizationId}/${name}/${version}/prepare-file-upload`,
       {
@@ -107,7 +117,7 @@ export class RegistryApi {
     isPublic,
     batchId,
     batchSize,
-  }: RegistryApiTypes.PushApiParams) {
+  }: PushApiParams) {
     const response = await this.request(
       `/${organizationId}/${name}/${version}`,
       {