+++ packages/core/src/rules/common/no-invalid-schema-examples.ts
@@ -1,9 +1,9 @@
 import { UserContext } from '../../walk';
 import { Oas3_1Schema } from '../../typings/openapi';
-import { validateExample } from '../utils';
+import { getAdditionalPropertiesOption, validateExample } from '../utils';
 
 export const NoInvalidSchemaExamples: any = (opts: any) => {
-  const disallowAdditionalProperties = opts.disallowAdditionalProperties ?? true;
+  const allowAdditionalProperties = getAdditionalPropertiesOption(opts) ?? false;
   return {
     Schema: {
       leave(schema: Oas3_1Schema, ctx: UserContext) {
@@ -14,12 +14,12 @@ export const NoInvalidSchemaExamples: any = (opts: any) => {
               schema,
               ctx.location.child(['examples', schema.examples.indexOf(example)]),
               ctx,
-              disallowAdditionalProperties
+              allowAdditionalProperties
             );
           }
         }
         if (schema.example) {
-          validateExample(schema.example, schema, ctx.location.child('example'), ctx, false);
+          validateExample(schema.example, schema, ctx.location.child('example'), ctx, true);
         }
       },
     },