+++ docs/changelog.md
@@ -4,6 +4,29 @@ tocMaxDepth: 2
 
 # Redocly CLI changelog
 
+
+## 1.0.0-beta.107 (2022-08-16)
+
+### Changes
+
+- Add `allowAdditionalProperties` to built-in rules and mark `disallowAdditionalProperties` as deprecated
+
+### Fixes
+
+- Introduced severity level `off` for Assertions
+
+
+## 1.0.0-beta.106 (2022-08-09)
+
+### Fixes
+
+- Now errors exit with return code `1`.
+
+### Changes
+
+- Renamed `lint` into `styleguide` in Redocly configuration.
+- Improved naming consistency.
+
 ## 1.0.0-beta.105 (2022-07-27)
 
 ### Fixes